+++ docker_daemon/CHANGELOG.md
@@ -1,7 +1,20 @@
 # CHANGELOG - docker_daemon
 
+1.6.0 / Unreleased
+==================
+### Changes
 
-1.5.0 / Unreleased
+* [IMPROVEMENT] Add custom tags to all service checks. See [#782][]
+* [IMPROVEMENT] Add docker memory soft limit metric. See [#760][]
+* [IMPROVEMENT] Add docker.containers.running.total & docker.containers.stopped.total metrics. See [#859][]
+
+1.5.1 / 2017-11-08
+==================
+### Changes
+
+* [BUGFIX] Fix lost kubernetes tags in 1.5.0. See [#817][]
+
+1.5.0 / 2017-10-10
 ==================
 ### Changes
 
@@ -100,4 +113,9 @@
 [#616]: https://github.com/DataDog/integrations-core/issues/616
 [#641]: https://github.com/DataDog/integrations-core/issues/641
 [#701]: https://github.com/DataDog/integrations-core/issues/701
+[#722]: https://github.com/DataDog/integrations-core/issues/722
+[#744]: https://github.com/DataDog/integrations-core/issues/744
+[#770]: https://github.com/DataDog/integrations-core/issues/770
+[#782]: https://github.com/DataDog/integrations-core/issues/782
+[#817]: https://github.com/DataDog/integrations-core/issues/817
 [@sophaskins]: https://github.com/sophaskins